+++ internal/commands/objectstorage/delete_test.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/UpCloudLtd/upcloud-go-api/v8/upcloud"
 	"github.com/UpCloudLtd/upcloud-go-api/v8/upcloud/request"
-	"github.com/gemalto/flume"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -56,6 +55,15 @@ func TestDeleteCommand(t *testing.T) {
 			flags: []string{"--delete-users", "--delete-policies"},
 			req:   request.DeleteManagedObjectStorageRequest{UUID: objectstorage.UUID},
 		},
+		{
+			name:  "delete with UUID using force",
+			arg:   objectstorage.UUID,
+			flags: []string{"--force"},
+			req: request.DeleteManagedObjectStorageRequest{
+				UUID:  objectstorage.UUID,
+				Force: true,
+			},
+		},
 	} {
 		t.Run(test.name, func(t *testing.T) {
 			mService := smock.Service{}
@@ -67,7 +75,7 @@ func TestDeleteCommand(t *testing.T) {
 			err := c.Cobra().Flags().Parse(test.flags)
 			assert.NoError(t, err)
 
-			_, err = c.(commands.MultipleArgumentCommand).Execute(commands.NewExecutor(conf, &mService, flume.New("test")), test.arg)
+			_, err = c.(commands.MultipleArgumentCommand).Execute(commands.NewExecutor(conf, &mService, conf.NewLogger("test")), test.arg)
 
 			if test.error != "" {
 				assert.EqualError(t, err, test.error)