+++ internal/mock/mock.go
@@ -88,7 +88,12 @@ func (m *Service) GetZones(context.Context) (*upcloud.Zones, error) {
 }
 
 // GetPriceZones implements service.Zones.GetPriceZones
-func (m *Service) GetPriceZones(context.Context) (*upcloud.PriceZones, error) {
+func (m *Service) GetPriceZones(context.Context) (*upcloud.PriceZones, error) { //nolint: staticcheck // Required to implement interface
+	return nil, nil
+}
+
+// GetPricingByZone implements service.Zones.GetPricingByZone
+func (m *Service) GetPricesByZone(context.Context) (*upcloud.PricesByZone, error) {
 	return nil, nil
 }
 
@@ -1312,6 +1317,22 @@ func (m *Service) DetachManagedObjectStorageUserPolicy(ctx context.Context, r *r
 	return nil
 }
 
+func (m *Service) CreateManagedObjectStoragePolicyVersion(ctx context.Context, r *request.CreateManagedObjectStoragePolicyVersionRequest) (*upcloud.ManagedObjectStoragePolicyVersion, error) {
+	return nil, nil
+}
+
+func (m *Service) GetManagedObjectStoragePolicyVersion(ctx context.Context, r *request.GetManagedObjectStoragePolicyVersionRequest) (*upcloud.ManagedObjectStoragePolicyVersion, error) {
+	return nil, nil
+}
+
+func (m *Service) GetManagedObjectStoragePolicyVersions(ctx context.Context, r *request.GetManagedObjectStoragePolicyVersionsRequest) ([]upcloud.ManagedObjectStoragePolicyVersion, error) {
+	return nil, nil
+}
+
+func (m *Service) DeleteManagedObjectStoragePolicyVersion(ctx context.Context, r *request.DeleteManagedObjectStoragePolicyVersionRequest) error {
+	return nil
+}
+
 func (m *Service) CreateManagedObjectStorageCustomDomain(ctx context.Context, r *request.CreateManagedObjectStorageCustomDomainRequest) error {
 	return m.Called(r).Error(0)
 }
@@ -1590,4 +1611,218 @@ func (m *Service) ExportAuditLog(_ context.Context, r *request.ExportAuditLogReq
 		return nil, args.Error(1)
 	}
 	return args[0].(io.ReadCloser), args.Error(1)
+}
+
+func (m *Service) GetFileStorages(ctx context.Context, r *request.GetFileStoragesRequest) ([]upcloud.FileStorage, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].([]upcloud.FileStorage), args.Error(1)
+}
+
+func (m *Service) CreateFileStorage(ctx context.Context, r *request.CreateFileStorageRequest) (*upcloud.FileStorage, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].(*upcloud.FileStorage), args.Error(1)
+}
+
+func (m *Service) GetFileStorage(ctx context.Context, r *request.GetFileStorageRequest) (*upcloud.FileStorage, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].(*upcloud.FileStorage), args.Error(1)
+}
+
+func (m *Service) ReplaceFileStorage(ctx context.Context, r *request.ReplaceFileStorageRequest) (*upcloud.FileStorage, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].(*upcloud.FileStorage), args.Error(1)
+}
+
+func (m *Service) ModifyFileStorage(ctx context.Context, r *request.ModifyFileStorageRequest) (*upcloud.FileStorage, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].(*upcloud.FileStorage), args.Error(1)
+}
+
+func (m *Service) DeleteFileStorage(ctx context.Context, r *request.DeleteFileStorageRequest) error {
+	args := m.Called(r)
+	return args.Error(0)
+}
+
+func (m *Service) WaitForFileStorageDeletion(ctx context.Context, r *request.WaitForFileStorageDeletionRequest) error {
+	args := m.Called(r)
+	return args.Error(0)
+}
+
+func (m *Service) WaitForFileStorageOperationalState(ctx context.Context, r *request.WaitForFileStorageOperationalStateRequest) (*upcloud.FileStorage, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].(*upcloud.FileStorage), args.Error(1)
+}
+
+func (m *Service) GetFileStorageCurrentState(ctx context.Context, r *request.GetFileStorageCurrentStateRequest) (*upcloud.FileStorage, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].(*upcloud.FileStorage), args.Error(1)
+}
+
+func (m *Service) GetFileStorageNetworks(ctx context.Context, r *request.GetFileStorageNetworksRequest) ([]upcloud.FileStorageNetwork, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].([]upcloud.FileStorageNetwork), args.Error(1)
+}
+
+func (m *Service) CreateFileStorageNetwork(ctx context.Context, r *request.CreateFileStorageNetworkRequest) (*upcloud.FileStorageNetwork, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].(*upcloud.FileStorageNetwork), args.Error(1)
+}
+
+func (m *Service) GetFileStorageNetwork(ctx context.Context, r *request.GetFileStorageNetworkRequest) (*upcloud.FileStorageNetwork, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].(*upcloud.FileStorageNetwork), args.Error(1)
+}
+
+func (m *Service) ModifyFileStorageNetwork(ctx context.Context, r *request.ModifyFileStorageNetworkRequest) (*upcloud.FileStorageNetwork, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].(*upcloud.FileStorageNetwork), args.Error(1)
+}
+
+func (m *Service) DeleteFileStorageNetwork(ctx context.Context, r *request.DeleteFileStorageNetworkRequest) error {
+	args := m.Called(r)
+	return args.Error(0)
+}
+
+func (m *Service) GetFileStorageShares(ctx context.Context, r *request.GetFileStorageSharesRequest) ([]upcloud.FileStorageShare, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].([]upcloud.FileStorageShare), args.Error(1)
+}
+
+func (m *Service) CreateFileStorageShare(ctx context.Context, r *request.CreateFileStorageShareRequest) (*upcloud.FileStorageShare, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].(*upcloud.FileStorageShare), args.Error(1)
+}
+
+func (m *Service) GetFileStorageShare(ctx context.Context, r *request.GetFileStorageShareRequest) (*upcloud.FileStorageShare, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].(*upcloud.FileStorageShare), args.Error(1)
+}
+
+func (m *Service) ModifyFileStorageShare(ctx context.Context, r *request.ModifyFileStorageShareRequest) (*upcloud.FileStorageShare, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].(*upcloud.FileStorageShare), args.Error(1)
+}
+
+func (m *Service) DeleteFileStorageShare(ctx context.Context, r *request.DeleteFileStorageShareRequest) error {
+	args := m.Called(r)
+	return args.Error(0)
+}
+
+func (m *Service) GetFileStorageShareACLs(ctx context.Context, r *request.GetFileStorageShareACLsRequest) ([]upcloud.FileStorageShareACL, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].([]upcloud.FileStorageShareACL), args.Error(1)
+}
+
+func (m *Service) CreateFileStorageShareACL(ctx context.Context, r *request.CreateFileStorageShareACLRequest) (*upcloud.FileStorageShareACL, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].(*upcloud.FileStorageShareACL), args.Error(1)
+}
+
+func (m *Service) GetFileStorageShareACL(ctx context.Context, r *request.GetFileStorageShareACLRequest) (*upcloud.FileStorageShareACL, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].(*upcloud.FileStorageShareACL), args.Error(1)
+}
+
+func (m *Service) ModifyFileStorageShareACL(ctx context.Context, r *request.ModifyFileStorageShareACLRequest) (*upcloud.FileStorageShareACL, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].(*upcloud.FileStorageShareACL), args.Error(1)
+}
+
+func (m *Service) DeleteFileStorageShareACL(ctx context.Context, r *request.DeleteFileStorageShareACLRequest) error {
+	args := m.Called(r)
+	return args.Error(0)
+}
+
+func (m *Service) GetFileStorageLabels(ctx context.Context, r *request.GetFileStorageLabelsRequest) ([]upcloud.Label, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].([]upcloud.Label), args.Error(1)
+}
+
+func (m *Service) CreateFileStorageLabel(ctx context.Context, r *request.CreateFileStorageLabelRequest) (*upcloud.Label, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].(*upcloud.Label), args.Error(1)
+}
+
+func (m *Service) GetFileStorageLabel(ctx context.Context, r *request.GetFileStorageLabelRequest) (*upcloud.Label, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].(*upcloud.Label), args.Error(1)
+}
+
+func (m *Service) ModifyFileStorageLabel(ctx context.Context, r *request.ModifyFileStorageLabelRequest) (*upcloud.Label, error) {
+	args := m.Called(r)
+	if args[0] == nil {
+		return nil, args.Error(1)
+	}
+	return args[0].(*upcloud.Label), args.Error(1)
+}
+
+func (m *Service) DeleteFileStorageLabel(ctx context.Context, r *request.DeleteFileStorageLabelRequest) error {
+	args := m.Called(r)
+	return args.Error(0)
 }